Update deposit function
Removed "require" that always fails, as the bond's expected balance of TIME should be: balanceBefore + payout

Depositing in the treasury an amount and profit would make the treasury mint amount - profit TIME to the bond. Then the DAO fee amount is used to send to the DAO. So the Bond's balance left is its balance before the deposit plus the payout to the bonder.